Tsuru now uses generic command line helper
This will be useful to implement Git helper
